+++ packages/xstate-graph/src/machineToBehavior.ts
@@ -1,0 +1,12 @@
+import { AnyStateMachine, StateFrom, EventFrom } from 'xstate';
+import { SimpleBehavior } from './types';
+
+export function machineToBehavior<TMachine extends AnyStateMachine>(
+  machine: TMachine
+): SimpleBehavior<StateFrom<TMachine>, EventFrom<TMachine>> {
+  return {
+    transition: (state, event) =>
+      machine.transition(state, event) as StateFrom<TMachine>,
+    initialState: machine.initialState as StateFrom<TMachine>
+  };
+}
