+++ packages/xstate-graph/src/shortestPaths.ts
@@ -14,16 +14,7 @@ import {
 } from './graph';
 import { getAdjacencyMap } from './adjacency';
 import { flatten } from 'xstate/src/utils';
-
-export function machineToBehavior<TMachine extends AnyStateMachine>(
-  machine: TMachine
-): SimpleBehavior<StateFrom<TMachine>, EventFrom<TMachine>> {
-  return {
-    transition: (state, event) =>
-      machine.transition(state, event) as StateFrom<TMachine>,
-    initialState: machine.initialState as StateFrom<TMachine>
-  };
-}
+import { machineToBehavior } from './machineToBehavior';
 
 export function getShortestPaths<TState, TEvent extends EventObject>(
   behavior: SimpleBehavior<TState, TEvent>,
